move access into its own feature + move access logic from controller to application service
