+++ src/Infrastructure/KafkaJanitor.RestApi/Features/Access/Application/AccessService.cs
@@ -7,20 +7,19 @@ using KafkaJanitor.RestApi.Features.ServiceAccounts.Infrastructure;
 using KafkaJanitor.RestApi.Features.Topics.Domain.Models;
 using KafkaJanitor.RestApi.Features.Vault;
 using KafkaJanitor.RestApi.Features.Vault.Model;
-using Microsoft.AspNetCore.Mvc;
 using Tika.RestClient.Features.ServiceAccounts.Models;
 
-namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
+namespace KafkaJanitor.RestApi.Features.Access.Application
 {
-    [Route(Routes.ACCESS_ROUTE)]
-    public class AccessController : ControllerBase
+    public class AccessService : IAccessService
     {
         private readonly IAccessControlListClient _accessControlListService;
         private readonly IServiceAccountClient _serviceAccountClient;
         private readonly IApiKeyClient _apiKeyClient;
         private readonly IVault _vault;
-        public AccessController(
-            IAccessControlListClient accessControlListService, 
+
+        public AccessService(
+            IAccessControlListClient accessControlListService,
             IServiceAccountClient serviceAccountClient,
             IApiKeyClient apiKeyClient,
             IVault vault
@@ -32,19 +31,14 @@ namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
             _vault = vault;
         }
 
-        [HttpPost("request")]
-        public async Task<IActionResult> RequestServiceAccount([FromBody] ServiceAccountRequestInput input)
+        public async Task ProvideAccess(
+            Capability cap,
+            string topicPrefix
+        )
         {
-            var cap = new Capability
-            {
-                Id = input.CapabilityId,
-                Name = input.CapabilityName,
-                RootId = input.CapabilityRootId
-            };
-
             ServiceAccount serviceAccount = null;
 
-            var doesServiceaccountExist = new Func<Task<bool>>(async () =>
+            Func<Task<bool>> doesServiceAccountExist = new Func<Task<bool>>(async () =>
             {
                 try
                 {
@@ -54,22 +48,25 @@ namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
                 {
                     return false;
                 }
+
                 return true;
             });
-            
-            var isExpectedAmountOfAclsInPlace = new Func<string,Task<bool>>(async (string serviceAccountId) =>
-            {
-                var acls = await _accessControlListService.GetAclsForServiceAccount(serviceAccountId);
-                return acls.Count() == 12;
-            });
-            
-            var isExpectedAmountOfApiKeysInPlace = new Func<string,Task<bool>>(async (string serviceAccountId) =>
-            {
-                var apiKeys = await _apiKeyClient.GetApiKeyPairsForServiceAccount(serviceAccountId);
-                return apiKeys.Any();
-            });
-            
-            if (!await doesServiceaccountExist())
+
+            Func<string, Task<bool>> isExpectedAmountOfAclsInPlace = new Func<string, Task<bool>>(
+                async (string serviceAccountId) =>
+                {
+                    var acls = await _accessControlListService.GetAclsForServiceAccount(serviceAccountId);
+                    return acls.Count() == 12;
+                });
+
+            Func<string, Task<bool>> isExpectedAmountOfApiKeysInPlace = new Func<string, Task<bool>>(
+                async (string serviceAccountId) =>
+                {
+                    var apiKeys = await _apiKeyClient.GetApiKeyPairsForServiceAccount(serviceAccountId);
+                    return apiKeys.Any();
+                });
+
+            if (!await doesServiceAccountExist())
             {
                 serviceAccount = await _serviceAccountClient.CreateServiceAccount(cap);
             }
@@ -80,7 +77,7 @@ namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
 
             if (!await isExpectedAmountOfAclsInPlace(serviceAccount.Id))
             {
-                await _accessControlListService.CreateAclsForServiceAccount(serviceAccount.Id, input.TopicPrefix);
+                await _accessControlListService.CreateAclsForServiceAccount(serviceAccount.Id, topicPrefix);
             }
 
             if (!await isExpectedAmountOfApiKeysInPlace(serviceAccount.Id))
@@ -88,25 +85,14 @@ namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
                 var apiKeyPair = await _apiKeyClient.CreateApiKeyPair(serviceAccount);
 
                 await _vault.AddApiCredentials(
-                    cap, 
+                    cap,
                     new ApiCredentials
                     {
                         Key = apiKeyPair.Key,
                         Secret = apiKeyPair.Secret
                     }
-                );   
+                );
             }
-
-            return Ok();
         }
-
-    }
-    
-    public class ServiceAccountRequestInput
-    {
-        public string CapabilityName { get; set; }
-        public string CapabilityId { get; set; }
-        public string CapabilityRootId { get; set; }
-        public string TopicPrefix { get; set; }
     }
 }
