+++ src/Infrastructure/KafkaJanitor.RestApi/Features/Access/Application/IAccessService.cs
@@ -1,0 +1,13 @@
+using System.Threading.Tasks;
+using KafkaJanitor.RestApi.Features.Topics.Domain.Models;
+
+namespace KafkaJanitor.RestApi.Features.Access.Application
+{
+    public interface IAccessService
+    {
+        Task ProvideAccess(
+            Capability cap,
+            string topicPrefix
+        );
+    }
+}
